backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
numpy>=1.26.0
pydantic==2.10.4
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.2.1
sentence-transformers==3.3.1
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
